fix(arquitetura): ajustar a arquitetura de diretorios do modelo
