+++ lib/octo_events/user/user.ex
